refactor(sysadmin): consolidate JST month helpers into report/util
The JST month boundary helpers (jstMonthStart / jstMonthStartOffset /
jstNextMonthStart / formatJstMonth) lived in
sysadmin/util.ts while the related day/week/arithmetic helpers
(truncateToJstDate / addDays / isoWeekStartJst / percentChange) lived
in report/util.ts. Callers needed two imports and two mental models
for "the JST utility".

Move the month-level helpers into report/util.ts next to the day-level
helpers they layer on top of. The report README already documented
them there; now the code matches the docs. Delete the now-empty
sysadmin/util.ts (bounds.ts from phase 1 remains for the asOfBounds
helper). Tests keep their existing file path but import from the
new location.

No behavioural change — 58 unit tests continue to pass; sysadmin
build remains clean.
